Bug250

+++ tests-ng/profile-dyninclude.sh
@@ -32,6 +32,7 @@ ddpath="${cur}/../"
 
 export PYTHONPATH="${ddpath}:${PYTHONPATH}"
 bin="python3 -m dotdrop.dotdrop"
+hash coverage 2>/dev/null && bin="coverage run -a --source=dotdrop -m dotdrop.dotdrop" || true
 
 echo "dotdrop path: ${ddpath}"
 echo "pythonpath: ${PYTHONPATH}"
@@ -90,6 +91,9 @@ dotfiles:
   f_def:
     dst: ${tmpd}/def
     src: def
+  f_ghi:
+    dst: '${tmpd}/{{@@ ghi @@}}'
+    src: ghi
 variables:
   mainvar: 'bad0'
   subvar: 'bad1'
@@ -100,8 +104,10 @@ profiles:
   subprofile:
     dotfiles:
     - f_abc
+    - f_ghi
     dynvariables:
       subdyn: 'echo subdyncontent'
+      ghi: 'echo ghi'
     variables:
       subvar: 'subcontent'
   subignore:
@@ -118,6 +124,7 @@ echo "{{@@ subdyn @@}}" >> ${tmps}/dotfiles/abc
 echo "{{@@ subvar @@}}" >> ${tmps}/dotfiles/abc
 echo "end" >> ${tmps}/dotfiles/abc
 #cat ${tmps}/dotfiles/abc
+echo "ghi content" > ${tmps}/dotfiles/ghi
 
 # install
 cd ${ddpath} | ${bin} install -f -c ${cfg} -p profile_1 --verbose
@@ -129,6 +136,7 @@ grep 'maindyncontent' ${tmpd}/abc >/dev/null || (echo "dynvariables 1 not resolv
 grep 'subcontent' ${tmpd}/abc >/dev/null || (echo "variables 2 not resolved" && exit 1)
 grep 'subdyncontent' ${tmpd}/abc >/dev/null || (echo "dynvariables 2 not resolved" && exit 1)
 #cat ${tmpd}/abc
+[ ! -e ${tmpd}/ghi ] && exit 1
 
 ## CLEANING
 rm -rf ${tmps} ${tmpd}
